requirements.txt
openai==1.12.0
rouge-score==0.1.2
python-dotenv==1.0.1